requirements.txt
requests==2.31.0
python-dotenv==1.0.1
pandas==2.2.0
openai==1.12.0
google-generativeai==0.4.0
httpx==0.27.0
websocket-client==1.7.0
pycryptodome==3.20.0
